# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.APICallLog:
    properties:
      channel:
        allOf:
        - $ref: '#/definitions/entity.Channel'
        description: 关联关系
      channel_id:
        type: integer
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      duration_ms:
        type: integer
      end_time:
        type: string
      error_message:
        type: string
      id:
        example: 1
        type: integer
      method:
        type: string
      processor:
        $ref: '#/definitions/entity.MessageProcessor'
      processor_id:
        type: integer
      request_body:
        type: string
      request_headers:
        $ref: '#/definitions/entity.JSONField'
      request_id:
        type: string
      response_body:
        type: string
      response_headers:
        $ref: '#/definitions/entity.JSONField'
      response_status:
        type: integer
      start_time:
        type: string
      success:
        type: boolean
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      url:
        type: string
    type: object
  entity.Bot:
    properties:
      avatar_url:
        type: string
      bot_name:
        type: string
      bot_type:
        type: string
      channels:
        description: 关联关系
        items:
          $ref: '#/definitions/entity.Channel'
        type: array
      config:
        $ref: '#/definitions/entity.JSONField'
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      created_by:
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      description:
        type: string
      id:
        example: 1
        type: integer
      status:
        $ref: '#/definitions/entity.Status'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  entity.Channel:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
      api_call_logs:
        items:
          $ref: '#/definitions/entity.APICallLog'
        type: array
      bot:
        allOf:
        - $ref: '#/definitions/entity.Bot'
        description: 关联关系
      bot_id:
        type: integer
      channel_name:
        type: string
      config:
        $ref: '#/definitions/entity.JSONField'
      connection_logs:
        items:
          $ref: '#/definitions/entity.ConnectionLog'
        type: array
      connection_status:
        $ref: '#/definitions/entity.ConnectionStatus'
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      id:
        example: 1
        type: integer
      last_connected_at:
        type: string
      messages:
        items:
          $ref: '#/definitions/entity.Message'
        type: array
      platform_type:
        type: string
      status:
        $ref: '#/definitions/entity.Status'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      webhook_path:
        type: string
    type: object
  entity.ConnectionLog:
    properties:
      channel:
        allOf:
        - $ref: '#/definitions/entity.Channel'
        description: 关联关系
      channel_id:
        type: integer
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      details:
        $ref: '#/definitions/entity.JSONField'
      error_code:
        type: string
      error_message:
        type: string
      event_type:
        type: string
      id:
        example: 1
        type: integer
      status:
        $ref: '#/definitions/entity.Status'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  entity.ConnectionStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      ConnectionStatusConnected: 已连接
      ConnectionStatusConnecting: 连接中
      ConnectionStatusDisconnected: 断开连接
      ConnectionStatusError: 连接错误
    x-enum-varnames:
    - ConnectionStatusDisconnected
    - ConnectionStatusConnected
    - ConnectionStatusConnecting
    - ConnectionStatusError
  entity.JSONField:
    additionalProperties: true
    type: object
  entity.Message:
    properties:
      backend_request_id:
        type: string
      channel:
        allOf:
        - $ref: '#/definitions/entity.Channel'
        description: 关联关系
      channel_id:
        type: integer
      content:
        type: string
      content_type:
        type: string
      conversation_id:
        type: string
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      direction:
        $ref: '#/definitions/entity.MessageDirection'
      error_message:
        type: string
      file_size:
        type: integer
      id:
        example: 1
        type: integer
      media_type:
        type: string
      media_url:
        type: string
      message_id:
        type: string
      message_status:
        $ref: '#/definitions/entity.MessageStatus'
      message_type:
        type: string
      parent_message:
        $ref: '#/definitions/entity.Message'
      parent_message_id:
        type: integer
      platform_message_id:
        type: string
      platform_timestamp:
        type: string
      processed_at:
        type: string
      raw_content:
        $ref: '#/definitions/entity.JSONField'
      received_at:
        type: string
      receiver_id:
        type: string
      receiver_name:
        type: string
      receiver_type:
        type: string
      replies:
        items:
          $ref: '#/definitions/entity.Message'
        type: array
      retry_count:
        type: integer
      sender_id:
        type: string
      sender_name:
        type: string
      sender_type:
        type: string
      sent_at:
        type: string
      unified_content:
        $ref: '#/definitions/entity.JSONField'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  entity.MessageDirection:
    enum:
    - 1
    - 2
    type: integer
    x-enum-comments:
      MessageDirectionInbound: 入站消息
      MessageDirectionOutbound: 出站消息
    x-enum-varnames:
    - MessageDirectionInbound
    - MessageDirectionOutbound
  entity.MessageProcessor:
    properties:
      api_call_logs:
        items:
          $ref: '#/definitions/entity.APICallLog'
        type: array
      config:
        $ref: '#/definitions/entity.JSONField'
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      id:
        example: 1
        type: integer
      priority:
        type: integer
      processor_name:
        type: string
      processor_type:
        type: string
      routing_rules:
        description: 关联关系
        items:
          $ref: '#/definitions/entity.MessageRoutingRule'
        type: array
      status:
        $ref: '#/definitions/entity.Status'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  entity.MessageRoutingRule:
    properties:
      bot_ids:
        $ref: '#/definitions/entity.JSONField'
      channel_ids:
        $ref: '#/definitions/entity.JSONField'
      condition_logic:
        type: string
      content_patterns:
        $ref: '#/definitions/entity.JSONField'
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      id:
        example: 1
        type: integer
      is_fallback:
        type: boolean
      message_processor:
        allOf:
        - $ref: '#/definitions/entity.MessageProcessor'
        description: 关联关系
      message_types:
        $ref: '#/definitions/entity.JSONField'
      platform_types:
        $ref: '#/definitions/entity.JSONField'
      priority:
        type: integer
      processor_id:
        type: integer
      route_type:
        $ref: '#/definitions/entity.RouteType'
      rule_description:
        type: string
      rule_name:
        type: string
      sender_patterns:
        $ref: '#/definitions/entity.JSONField'
      status:
        $ref: '#/definitions/entity.Status'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  entity.MessageStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-comments:
      MessageStatusExpired: 已过期
      MessageStatusFailed: 处理失败
      MessageStatusPending: 待处理
      MessageStatusProcessed: 已处理
      MessageStatusProcessing: 处理中
      MessageStatusSent: 已发送
    x-enum-varnames:
    - MessageStatusPending
    - MessageStatusProcessing
    - MessageStatusProcessed
    - MessageStatusSent
    - MessageStatusFailed
    - MessageStatusExpired
...
  entity.RouteType:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-comments:
      RouteTypeBroadcast: 广播路由
      RouteTypeConditional: 条件路由
      RouteTypeDirect: 直接路由
      RouteTypeForward: 转发路由
    x-enum-varnames:
    - RouteTypeDirect
    - RouteTypeForward
    - RouteTypeBroadcast
    - RouteTypeConditional
  entity.Status:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      StatusActive: 激活/启用
      StatusDeleted: 已删除
      StatusInactive: 未激活/禁用
      StatusSuspended: 暂停
    x-enum-varnames:
    - StatusInactive
    - StatusActive
    - StatusDeleted
    - StatusSuspended
...
  usecase.APICallLogListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.APICallLog'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  usecase.APICallStatsResult:
    properties:
      avg_duration_ms:
        type: number
      failed_calls:
        type: integer
      success_calls:
        type: integer
      success_rate:
        type: number
      total_calls:
        type: integer
    type: object
...
  usecase.ConnectionLogListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.ConnectionLog'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  v1.StandardResponse:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 统一消息机器人管理平台API
  title: OmniBotGo API
  version: "1.0"
paths:
...
  /api/v1/logs/api-calls:
    get:
      consumes:
      - application/json
      description: 获取API调用日志记录，支持分页和过滤
      parameters:
      - description: HTTP方法
        in: query
        name: method
        type: string
      - description: 请求路径
        in: query
        name: path
        type: string
      - description: 状态码
        in: query
        name: status_code
        type: integer
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      - default: 1
        description: 页码，默认1
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量，默认20
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/usecase.APICallLogListResult'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 获取API调用日志列表
      tags:
      - logs
  /api/v1/logs/api-calls/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取指定API调用日志的详细信息
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.APICallLog'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 日志不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 获取API调用日志详情
      tags:
      - logs
  /api/v1/logs/api-calls/stats:
    get:
      consumes:
      - application/json
      description: 获取API调用的统计信息，包括成功率、响应时间等
      parameters:
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      - description: 分组方式
        enum:
        - hour
        - day
        - week
        - month
        in: query
        name: group_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/usecase.APICallStatsResult'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 获取API调用统计
      tags:
      - logs
  /api/v1/logs/connections:
    get:
      consumes:
      - application/json
      description: 获取平台连接日志记录，支持分页和过滤
      parameters:
      - description: 通道ID
        in: query
        name: channel_id
        type: integer
      - description: 连接状态
        in: query
        name: status
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      - default: 1
        description: 页码，默认1
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量，默认20
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/usecase.ConnectionLogListResult'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 获取连接日志列表
      tags:
      - logs
  /api/v1/logs/connections/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取指定连接日志的详细信息
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.ConnectionLog'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 日志不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 获取连接日志详情
      tags:
      - logs
...
swagger: "2.0"